chore: renamed module to `mst` and modified relative paths to absolute
