+++ src/merkle_sum_tree/mod.rs
@@ -1,5 +1,5 @@
 mod entry;
-mod merkle_sum_tree_lib;
+mod mst;
 mod utils;
 mod tests;
 use halo2_proofs::halo2curves::bn256::{Fr as Fp};
@@ -20,7 +20,7 @@ pub struct Node {
 }
 
 pub use entry::Entry;
-pub use merkle_sum_tree_lib::MerkleSumTree;
+pub use mst::MerkleSumTree;
 pub use utils::{big_intify_username, big_int_to_fp};
