+++ src/merkle_sum_tree/utils/index_of.rs
@@ -1,5 +1,4 @@
 use num_bigint::BigInt;
-
 use crate::merkle_sum_tree::{Entry, Node};
 
 pub fn index_of(username: &str, balance: BigInt, nodes: &[Vec<Node>]) -> Option<usize> {
