+++ src/merkle_sum_tree/utils/mod.rs
@@ -8,12 +8,11 @@ mod proof_verification;
 mod operation_helpers;
 
 pub use build_tree::build_merkle_tree_from_entries;
-pub use create_middle_node::create_middle_node;
 pub use create_proof::create_proof;
 pub use csv_parser::parse_csv_to_entries;
 pub use hash::poseidon;
 pub use index_of::index_of;
 pub use proof_verification::verify_proof;
-pub use operation_helpers::{big_intify_username, big_int_to_fp};
+pub use operation_helpers::*;
